add Key class and refactor so Shift inherits Key
